downscaler/requirements.txt
ffmpeg-python
watchdog
openai-whisper
ultralytics
requests
numpy